+++ features/defi-wrapper/timelock.ts
@@ -21,6 +21,30 @@ import {
   getTimeLockContract,
   TimeLockContract,
 } from 'contracts/defi-wrapper/index.js';
+import { DashboardAbi } from 'abi/Dashboard.js';
+import {
+  TimeLockAbi,
+  DistributorAbi,
+  WithdrawalQueueAbi,
+  StvStETHPoolAbi,
+  StvPoolAbi,
+  OssifiableProxyAbi,
+  GenericStrategyAbi,
+  MellowStrategyAbi,
+} from 'abi/defi-wrapper/index.js';
+import { getPublicClient } from 'providers/wallet.js';
+
+export const targetContractTypeToAbi = {
+  proxy: OssifiableProxyAbi,
+  dashboard: DashboardAbi,
+  withdrawalQueue: WithdrawalQueueAbi,
+  distributor: DistributorAbi,
+  timelock: TimeLockAbi,
+  stvPool: StvPoolAbi,
+  stvStethPool: StvStETHPoolAbi,
+  genericStrategy: GenericStrategyAbi,
+  mellowStrategy: MellowStrategyAbi,
+} as const;
 
 // Common constants
 export const DEFAULT_PREDECESSOR = zeroHash;
@@ -61,6 +85,33 @@ export const SALT_OPTION = [
   DEFAULT_SALT,
 ] as const;
 
+export const PREDECESSOR_OPTION = [
+  '-p, --predecessor <predecessor>',
+  'predecessor operation ID (bytes32 hex, default: 0x0)',
+  stringToHash,
+  DEFAULT_PREDECESSOR,
+] as const;
+
+export const SKIP_SIMULATION_OPTION = [
+  '--skip-simulation',
+  'skip simulation of operations before proposing/executing (not recommended)',
+] as const;
+
+export const TARGET_CONTRACT_TYPE_ARGUMENT = [
+  '[targetContractType]',
+  'target contract type for predefined operations (proxy, dashboard, withdrawalQueue, distributor, timelock, stvPool, stvStethPool, genericStrategy, mellowStrategy)',
+  (input: string) => {
+    if (targetContractTypeToAbi.hasOwnProperty(input)) {
+      return input as keyof typeof targetContractTypeToAbi;
+    }
+    throw new Error(
+      `Invalid target contract type: ${input}. Valid types are: ${Object.keys(
+        targetContractTypeToAbi,
+      ).join(', ')}`,
+    );
+  },
+] as const;
+
 // Helper function to get timelock from argument or prompt user
 export const getPromptTimelock = async (
   argAddress: Address | undefined,
@@ -97,6 +148,13 @@ export const waitTimeTo = (timestamp: bigint) => {
   return timestamp > now ? timestamp - now : 0n;
 };
 
+// transient type to migrate args to object
+type TimelockOperationOptions = {
+  predecessor?: Hex;
+  value?: bigint;
+  skipSimulation?: boolean;
+};
+
 // Helper function for propose operations
 export const proposeOperation = async (
   timelock: Address,
@@ -105,9 +163,13 @@ export const proposeOperation = async (
   salt: Hex,
   functionName: string,
   confirmationMessage: string,
-  value = 0n,
-  predecessor: Hex = DEFAULT_PREDECESSOR,
+  options: TimelockOperationOptions = {},
 ): Promise<Hex> => {
+  const {
+    predecessor = DEFAULT_PREDECESSOR,
+    value = 0n,
+    skipSimulation = false,
+  } = options;
   const timelockContract = await getTimeLockContract(timelock);
   const minDelay = await callReadMethodSilent({
     contract: timelockContract,
@@ -130,6 +192,31 @@ export const proposeOperation = async (
   logInfo(`  Salt: ${salt}`);
   logInfo(`  Function: ${functionName}`);
   logInfo(`  Min delay: ${minDelay} seconds`);
+
+  if (!skipSimulation) {
+    // Simulate the operation to check for errors before proposing
+    const client = await getPublicClient();
+    const { results } = await client.simulateCalls({
+      account: timelock,
+      calls: [
+        {
+          to: target,
+          data,
+          value,
+        },
+      ],
+    });
+    if (results[0].error) {
+      logInfo(
+        `❌ Simulation failed for the proposed operation. Error: ${results[0].error.message}`,
+      );
+      throw new Error('Operation simulation failed. See logs for details.');
+    } else {
+      logInfo(
+        '  ✅ Successfully simulated the proposed operation. No errors detected.',
+      );
+    }
+  }
 
   const confirm = await confirmOperation(confirmationMessage);
   if (!confirm) {
@@ -181,9 +268,9 @@ export const executeOperation = async (
   salt: Hex,
   functionName: string,
   confirmationMessage: string,
-  value = 0n,
-  predecessor: Hex = DEFAULT_PREDECESSOR,
+  options: TimelockOperationOptions = {},
 ): Promise<void> => {
+  const { predecessor = DEFAULT_PREDECESSOR, value = 0n } = options;
   const timelockContract = await getTimeLockContract(timelock);
 
   const operationId = await callReadMethodSilent({